reconstruction registry center and config center
