+++ pkg/adapter/events/configcenter.go
