+++ pkg/adapter/registrycenter/nacos/client_test.go
