+++ pkg/adapter/zookeeper/pathcache_test.go
@@ -1,31 +1,22 @@
 package zookeeper
 
-import (
-	"fmt"
-	"testing"
-	"time"
+// func Test_watch(t *testing.T) {
+// 	conn, _, err := zk.Connect(ZkServers, time.Duration(15*time.Hour))
+// 	if err != nil {
+// 		fmt.Printf("connect to zookeeper has an err: %v\n", err)
+// 		return
+// 	}
 
-	"github.com/samuel/go-zookeeper/zk"
-	"github.com/stretchr/testify/assert"
-)
+// 	//go func() {
+// 	p, err := newPathCache(conn, DubboRootPath, "MOCK", true)
+// 	if err != nil {
+// 		fmt.Printf("Create a new pathcache for [%s] has an err: %v\n", DubboRootPath, err)
+// 		assert.EqualError(t, err, "zk: node does not exist")
+// 		return
+// 	}
+// 	fmt.Printf("Created a path cache : %s\n", p.path)
+// 	//}()
 
-func Test_watch(t *testing.T) {
-	conn, _, err := zk.Connect(ZkServers, time.Duration(15*time.Hour))
-	if err != nil {
-		fmt.Printf("connect to zookeeper has an err: %v\n", err)
-		return
-	}
+// 	time.Sleep(30 * time.Minute)
 
-	//go func() {
-	p, err := newPathCache(conn, DubboRootPath, "MOCK", true)
-	if err != nil {
-		fmt.Printf("Create a new pathcache for [%s] has an err: %v\n", DubboRootPath, err)
-		assert.EqualError(t, err, "zk: node does not exist")
-		return
-	}
-	fmt.Printf("Created a path cache : %s\n", p.path)
-	//}()
-
-	time.Sleep(30 * time.Minute)
-
-}
+// }
